+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Added debug logging for xcode-tools-bridge proxy tool calls to aid diagnosis of bridge-related issues ([#252](https://github.com/getsentry/XcodeBuildMCP/issues/252)).
+
+### Fixed
+
+- Fixed `jsonSchemaToZod` converter to apply `default` values from remote tool JSON schemas, preventing potential argument loss when proxying bridge tools ([#252](https://github.com/getsentry/XcodeBuildMCP/issues/252)).
+
 ## [2.3.2]
 
 ### Fixed